refactor: merge `messages/transmit.py` module
